minor tweaks, restructuring of contract_instance
