+++ examples/anvil/examples/deploy_contract_anvil.rs
@@ -9,7 +9,7 @@ use eyre::Result;
 // Codegen from embedded Solidity code and precompiled bytecode.
 sol! {
     #[allow(missing_docs)]
-    // solc v0.8.26; solc a.sol --via-ir --optimize --bin
+    // solc v0.8.26; solc Counter.sol --via-ir --optimize --bin
     #[sol(rpc, bytecode="6080806040523460135760df908160198239f35b600080fdfe6080806040526004361015601257600080fd5b60003560e01c9081633fb5c1cb1460925781638381f58a146079575063d09de08a14603c57600080fd5b3460745760003660031901126074576000546000198114605e57600101600055005b634e487b7160e01b600052601160045260246000fd5b600080fd5b3460745760003660031901126074576020906000548152f35b34607457602036600319011260745760043560005500fea2646970667358221220e978270883b7baed10810c4079c941512e93a7ba1cd1108c781d4bc738d9090564736f6c634300081a0033")]
     contract Counter {
         uint256 public number;
@@ -41,7 +41,7 @@ async fn main() -> Result<()> {
 
     println!("Anvil running at `{}`", anvil.endpoint());
 
-    // Deploy the contract.
+    // Deploy the `Counter` contract.
     let contract = Counter::deploy(&provider).await?;
 
     println!("Deployed contract at address: {}", contract.address());
