+++ examples/subscriptions/examples/poll_logs.rs
@@ -9,7 +9,7 @@ use eyre::Result;
 use futures_util::StreamExt;
 
 // Codegen from embedded Solidity code and precompiled bytecode.
-// solc v0.8.26; solc a.sol --via-ir --optimize --bin
+// solc v0.8.26; solc Counter.sol --via-ir --optimize --bin
 sol!(
     #[allow(missing_docs)]
     #[sol(rpc, bytecode = "6080806040523460195760008055610155908161001f8239f35b600080fdfe6080604052600436101561001257600080fd5b60003560e01c80632baeceb7146100d057806361bc221a146100b25763d09de08a1461003d57600080fd5b346100ad5760003660031901126100ad57600054600181019060006001831291129080158216911516176100975780600055337ff6d1d8d205b41f9fb9549900a8dba5d669d68117a3a2b88c1ebc61163e8117ba600080a3005b634e487b7160e01b600052601160045260246000fd5b600080fd5b346100ad5760003660031901126100ad576020600054604051908152f35b346100ad5760003660031901126100ad5760005460001981019081136001166100975780600055337fdc69c403b972fc566a14058b3b18e1513da476de6ac475716e489fae0cbe4a26600080a300fea26469706673582212200d333e08e1230b0b9919825888e587a45c68e2aa2f7f58752712491e2201da9c64736f6c634300081a0033")]
